# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/settings/organization": {
      "get": {
        "operationId": "SettingsController_getOrganization",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Get organization profile",
        "tags": [
          "Settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateOrganization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Acme Pvt Ltd"
                  },
                  "legal_name": {
                    "type": "string",
                    "example": "Acme Private Limited"
                  },
                  "gstin": {
                    "type": "string",
                    "example": "29ABCDE1234F1Z5"
                  },
                  "pan": {
                    "type": "string",
                    "example": "ABCDE1234F"
                  },
                  "cin": {
                    "type": "string",
                    "example": "U72200KA2010PTC053718"
                  },
                  "address": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "phone": {
                    "type": "string",
                    "example": "+919876543210"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "contact@acme.com"
                  },
                  "logo_url": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://cdn.kontafy.com/logos/abc.png"
                  },
                  "fiscal_year_start": {
                    "type": "number",
                    "example": 4
                  },
                  "business_type": {
                    "type": "string",
                    "example": "Private Limited"
                  },
                  "industry": {
                    "type": "string",
                    "example": "Technology"
                  },
                  "currency": {
                    "type": "string",
                    "example": "INR"
                  },
                  "tan": {
                    "type": "string",
                    "example": "BLRA12345B"
                  },
                  "website": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://acme.com"
                  },
                  "date_of_incorporation": {
                    "type": "string",
                    "format": "date",
                    "example": "2010-04-15"
                  },
                  "gst_registration_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2017-07-01"
                  },
                  "books_begin_from": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-04-01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Update organization profile",
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/organization/logo": {
      "post": {
        "operationId": "SettingsController_uploadLogo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "data_url"
                ],
                "properties": {
                  "data_url": {
                    "type": "string",
                    "description": "Image as a base64 data URL",
                    "example": "data:image/png;base64,iVBORw0KGgo..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Upload organization logo (data URL)",
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/users": {
      "get": {
        "operationId": "SettingsController_listMembers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "List organization members with roles",
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/users/invite": {
      "post": {
        "operationId": "SettingsController_inviteUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "newuser@example.com"
                  },
                  "role": {
                    "type": "string",
                    "example": "accountant"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Invite user by email",
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/users/{id}/role": {
      "patch": {
        "operationId": "SettingsController_updateMemberRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "example": "admin"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Change member role",
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/users/{id}": {
      "delete": {
        "operationId": "SettingsController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Remove member from organization",
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/invoice-config": {
      "get": {
        "operationId": "SettingsController_getInvoiceConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Get invoice numbering, prefix, terms, bank details",
        "tags": [
          "Settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateInvoiceConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "invoice_prefix": {
                    "type": "string",
                    "example": "INV-"
                  },
                  "next_invoice_number": {
                    "type": "number",
                    "example": 1001
                  },
                  "default_payment_terms": {
                    "type": "number",
                    "example": 30
                  },
                  "default_terms_conditions": {
                    "type": "string",
                    "example": "Payment due within 30 days."
                  },
                  "default_notes": {
                    "type": "string",
                    "example": "Thank you for your business."
                  },
                  "bank_name": {
                    "type": "string",
                    "example": "HDFC Bank"
                  },
                  "bank_account_number": {
                    "type": "string",
                    "example": "50100123456789"
                  },
                  "bank_ifsc": {
                    "type": "string",
                    "example": "HDFC0001234"
                  },
                  "bank_branch": {
                    "type": "string",
                    "example": "MG Road, Bangalore"
                  },
                  "bank_accounts": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "bank_name",
                        "account_number",
                        "ifsc"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "bank_name": {
                          "type": "string",
                          "example": "HDFC Bank"
                        },
                        "account_name": {
                          "type": "string",
                          "example": "Acme Pvt Ltd"
                        },
                        "account_number": {
                          "type": "string",
                          "example": "50100123456789"
                        },
                        "ifsc": {
                          "type": "string",
                          "example": "HDFC0001234"
                        },
                        "branch": {
                          "type": "string",
                          "example": "MG Road, Bangalore"
                        },
                        "account_type": {
                          "type": "string",
                          "example": "current"
                        },
                        "upi_id": {
                          "type": "string",
                          "example": "acme@hdfcbank"
                        },
                        "swift_code": {
                          "type": "string",
                          "example": "HDFCINBB"
                        },
                        "is_primary": {
                          "type": "boolean",
                          "example": true
                        },
                        "show_full_number": {
                          "type": "boolean",
                          "example": false
                        },
                        "opening_balance": {
                          "type": "number",
                          "example": 0
                        },
                        "opening_dr_cr": {
                          "type": "string",
                          "enum": [
                            "Dr",
                            "Cr"
                          ],
                          "example": "Dr"
                        },
                        "opening_date": {
                          "type": "string",
                          "format": "date",
                          "example": "2025-04-01"
                        },
                        "account_id": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Update invoice configuration",
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/capital": {
      "get": {
        "operationId": "SettingsController_getCapitalStructure",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Get authorized / paid-up capital structure",
        "tags": [
          "Settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateCapitalStructure",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "authorized_capital": {
                    "type": "number",
                    "example": 1000000
                  },
                  "authorized_shares": {
                    "type": "number",
                    "example": 100000
                  },
                  "face_value": {
                    "type": "number",
                    "example": 10
                  },
                  "paid_up_capital": {
                    "type": "number",
                    "example": 500000
                  },
                  "issued_shares": {
                    "type": "number",
                    "example": 50000
                  },
                  "share_type": {
                    "type": "string",
                    "example": "equity"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Update authorized / paid-up capital structure",
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/directors": {
      "get": {
        "operationId": "SettingsController_getDirectors",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "List directors / signatories",
        "tags": [
          "Settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateDirectors",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "directors"
                ],
                "properties": {
                  "directors": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "org-id": []
          }
        ],
        "summary": "Replace the directors list (whole-array PATCH)",
        "tags": [
          "Settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kontafy API",
    "description": "REST API for Kontafy — cloud-native accounting for Indian businesses. All paths are versioned: prefix every request with `/v1`. Authenticated routes expect `Authorization: Bearer <access_token>` and most multi-tenant routes also expect `X-Org-Id: <organization id>`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.kontafy.com/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:5002/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /v1/auth/login"
      },
      "org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Org-Id",
        "description": "Organization id for multi-tenant routes"
      }
    },
...
  }
}